[api-refactor] Move type creation into LLVMType
