+++ src/main/kotlin/dev/supergrecko/kllvm/core/type/LLVMType.kt
@@ -1,5 +1,8 @@
 package dev.supergrecko.kllvm.core.type
 
+import dev.supergrecko.kllvm.utils.toInt
+import org.bytedeco.javacpp.PointerPointer
+import org.bytedeco.llvm.LLVM.LLVMContextRef
 import org.bytedeco.llvm.LLVM.LLVMTypeRef
 import org.bytedeco.llvm.global.LLVM
 
@@ -54,5 +57,76 @@ public open class LLVMType internal constructor(internal val llvmType: LLVMTypeR
         LLVM_I64_TYPE,
         LLVM_I128_TYPE,
         LLVM_INT_TYPE
+    }
+
+    companion object {
+        /**
+         * Create a type in a context and a known size.
+         *
+         * @param kind Integer kind to create
+         * @param size Context size for [LLVMType.IntegerTypeKinds.LLVM_INT_TYPE]
+         * @param context The context to use, default to global
+         *
+         * @throws IllegalArgumentException If wanted size is less than 0 or larger than 2^23-1
+         */
+        @JvmStatic
+        public fun makeInteger(kind: IntegerTypeKinds, size: Int = 0, context: LLVMContextRef = LLVM.LLVMGetGlobalContext()): LLVMIntegerType {
+            val type = when (kind) {
+                IntegerTypeKinds.LLVM_I1_TYPE -> LLVM.LLVMInt1TypeInContext(context)
+                IntegerTypeKinds.LLVM_I8_TYPE -> LLVM.LLVMInt8TypeInContext(context)
+                IntegerTypeKinds.LLVM_I16_TYPE -> LLVM.LLVMInt16TypeInContext(context)
+                IntegerTypeKinds.LLVM_I32_TYPE -> LLVM.LLVMInt32TypeInContext(context)
+                IntegerTypeKinds.LLVM_I64_TYPE -> LLVM.LLVMInt64TypeInContext(context)
+                IntegerTypeKinds.LLVM_I128_TYPE -> LLVM.LLVMInt128TypeInContext(context)
+                IntegerTypeKinds.LLVM_INT_TYPE -> {
+                    require(size in 1..8388606) { "LLVM only supports integers of 2^23-1 bits size" }
+
+                    LLVM.LLVMIntTypeInContext(context, size)
+                }
+            }
+
+            return LLVMIntegerType(type)
+        }
+
+        /**
+         * Create a float type in the given context
+         *
+         * @param kind Float type to create
+         * @param context The context to use, default to global
+         */
+        @JvmStatic
+        public fun makeFloat(kind: FloatTypeKinds, context: LLVMContextRef = LLVM.LLVMGetGlobalContext()): LLVMFloatType {
+            val type = when (kind) {
+                FloatTypeKinds.LLVM_HALF_TYPE -> LLVM.LLVMHalfTypeInContext(context)
+                FloatTypeKinds.LLVM_FLOAT_TYPE -> LLVM.LLVMFloatTypeInContext(context)
+                FloatTypeKinds.LLVM_DOUBLE_TYPE -> LLVM.LLVMDoubleTypeInContext(context)
+                FloatTypeKinds.LLVM_X86FP80_TYPE -> LLVM.LLVMX86FP80TypeInContext(context)
+                FloatTypeKinds.LLVM_FP128_TYPE -> LLVM.LLVMFP128TypeInContext(context)
+                FloatTypeKinds.LLVM_PPCFP128_TYPE -> LLVM.LLVMPPCFP128TypeInContext(context)
+            }
+
+            return LLVMFloatType(type)
+        }
+
+        /**
+         * Create a function type
+         *
+         * Argument count is automatically calculated from [paramTypes].
+         *
+         * @param returnType Expected return type
+         * @param paramTypes List of parameter types
+         * @param isVariadic Is the function variadic?
+         */
+        @JvmStatic
+        public fun makeFunction(returnType: LLVMType, paramTypes: List<LLVMType>, isVariadic: Boolean): LLVMFunctionType {
+            val types = paramTypes.map { it.llvmType }
+            val array = ArrayList(types).toTypedArray()
+
+            val ptr = PointerPointer(*array)
+
+            val type = LLVM.LLVMFunctionType(returnType.llvmType, ptr, paramTypes.size, isVariadic.toInt())
+
+            return LLVMFunctionType(type, returnType, paramTypes)
+        }
     }
 }
