+++ src/test/kotlin/dev/supergrecko/kllvm/core/type/LLVMFunctionTypeTest.kt
@@ -8,9 +8,8 @@ import kotlin.test.assertEquals
 class LLVMFunctionTypeTest {
     @Test
     fun `creation of zero arg type works`() {
-        val ret = LLVMIntegerType.type(64)
-
-        val fn = LLVMFunctionType.type(ret, listOf(), false)
+        val ret = LLVMType.makeInteger(LLVMType.IntegerTypeKinds.LLVM_I64_TYPE)
+        val fn = LLVMType.makeFunction(ret, listOf(), false)
 
         assertEquals(fn.getParameterCount(), 0)
         assertEquals(fn.getReturnType().llvmType, ret.llvmType)
@@ -18,30 +17,27 @@ class LLVMFunctionTypeTest {
 
     @Test
     fun `variadic arguments work`() {
-        val ret = LLVMIntegerType.type(64)
-        val arg = LLVMFloatType.type(LLVMType.FloatTypeKinds.LLVM_FLOAT_TYPE)
-
-        val fn = LLVMFunctionType.type(ret, listOf(arg), true)
+        val ret = LLVMType.makeInteger(LLVMType.IntegerTypeKinds.LLVM_I64_TYPE)
+        val arg = LLVMType.makeFloat(LLVMType.FloatTypeKinds.LLVM_FLOAT_TYPE)
+        val fn = LLVMType.makeFunction(ret, listOf(arg), true)
 
         assertEquals(fn.isVariadic(), true)
     }
 
     @Test
     fun `test variadic wrapper works`() {
-        val ret = LLVMIntegerType.type(64)
-        val arg = LLVMFloatType.type(LLVMType.FloatTypeKinds.LLVM_FLOAT_TYPE)
-
-        val fn = LLVMFunctionType.type(ret, listOf(arg), true)
+        val ret = LLVMType.makeInteger(LLVMType.IntegerTypeKinds.LLVM_I64_TYPE)
+        val arg = LLVMType.makeFloat(LLVMType.FloatTypeKinds.LLVM_FLOAT_TYPE)
+        val fn = LLVMType.makeFunction(ret, listOf(arg), true)
 
         assertEquals(LLVM.LLVMIsFunctionVarArg(fn.llvmType).toBoolean(), fn.isVariadic())
     }
 
     @Test
     fun `test parameter count wrapper works`() {
-        val ret = LLVMIntegerType.type(64)
-        val arg = LLVMFloatType.type(LLVMType.FloatTypeKinds.LLVM_FLOAT_TYPE)
-
-        val fn = LLVMFunctionType.type(ret, listOf(arg), true)
+        val ret = LLVMType.makeInteger(LLVMType.IntegerTypeKinds.LLVM_I64_TYPE)
+        val arg = LLVMType.makeFloat(LLVMType.FloatTypeKinds.LLVM_FLOAT_TYPE)
+        val fn = LLVMType.makeFunction(ret, listOf(arg), true)
 
         assertEquals(LLVM.LLVMCountParamTypes(fn.llvmType), fn.getParameterCount())
     }
